% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdi_functions.R
\name{calculate_daily_rainfall}
\alias{calculate_daily_rainfall}
\title{Aggregate rainfall values to daily time steps}
\usage{
calculate_daily_rainfall(
  dat,
  datatype = c("guess", "aws", "synoptic"),
  crop = FALSE,
  on.error = c("stop", "null")
)
}
\arguments{
\item{dat}{A data set of sub-daily weather records for either an AWS or a
Synoptic source.}

\item{datatype}{Source type of data. Default is to guess based on data
values.}

\item{crop}{If \code{TRUE} (default), discard the rainfall for an initial
part day.}

\item{on.error}{Action to perform if there is any interruption in the time
series. If \code{'stop'}, the function will stop with an error message. If
\code{'null'}, the function will return NULL.}
}
\value{
A data frame with columns: station (if present in the input data),
  date, precipdaily.
}
\description{
This is a helper function called by other functions when
calculating fire-related variables. It aggregates sub-daily rainfall data to
daily values, taking into account the different conventions used for AWS and
Syoptic data sources. Each value for each day is the total rainfall recorded
from 09:01 that day to 09:00 the following day. If there are any missing days
in the time series, the function will either stop with an error or return
\code{NULL} depending on the value of the \code{on.error} argument.
}
