% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdi_functions.R
\name{calculate_effective_rainfall}
\alias{calculate_effective_rainfall}
\title{Calculate effective rainfall values from daily values}
\usage{
calculate_effective_rainfall(precipdaily, start.balance = 0)
}
\arguments{
\item{precipdaily}{Vector of daily precipitation values.}

\item{start.balance}{Starting balance. Should be a value between 0 and 5
(default is 0). A value outside this range will be silently set to 0 or 5.}
}
\value{
A vector of effective rainfall values.
}
\description{
For KBDI calculation (Finkele et al. 2006), the first 5mm of rain
do not count towards evapo-transpiration. If it rains less than 5mm over
several days, a running balance is kept. Once this balance exceeds 5mm, the
first 5mm is removed for canopy interception/runoff and the remainder
constitutes effective rainfall. Any day without rain resets the balance to
zero.
}
