% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdi_functions.R
\name{calculate_kbdi}
\alias{calculate_kbdi}
\title{Calculate daily KBDI values from data for a single weather station}
\usage{
calculate_kbdi(dat.daily, average.rainfall, assume.order = FALSE)
}
\arguments{
\item{dat.daily}{A data frame of daily weather data. This should be for a
single weather station. If a 'station' column is present this will be
checked, otherwise it is assumed. Must have columns: date,
precipdaily, tmaxdaily. Optionally, it can also have column: kbdi.
Additional columns are permitted but will be ignored. See Description for
details.}

\item{average.rainfall}{Reference value for average annual rainfall.}

\item{assume.order}{If \code{TRUE} the daily data are assumed to be in
calendar order and form an uninterrupted time series. If \code{FALSE}
(default) this will be checked. This option provides for more efficient
processing as part of a pipeline, but should be left at the default when
the function is called directly.}
}
\value{
A data frame with columns: date, kbdi.
}
\description{
Calculates daily values for the Keetch-Byram drought index. Following Finkele
et al. (2006), KBDI is calculated as yesterday's KBDI plus
evapo-transpiration minus effective rainfall. Input is a data frame of daily
weather data. This must have columns: date, precipdaily,
tmaxdaily. Optionally, it can also have a column: kbdi. If present,
calculations will be done for the most recent block of records with missing
values, with preceding records being used to initialize the KBDI time series.
If a kbdi column is absent, or present but with all missing values,
calculations will be performed for all records using the default
initialization in which KBDI is set to its maximum value (203.2).
}
\details{
Dealing with missing values: Nny missing daily precipitation values are set
to zero and a message is issued, Missing daily temperature values are left
as-is. The value of KBDI depends on the previous day's value. A missing value
in the daily time series for temperature leads to a missing value for KBDI,
which would then propagate to the next and all subsequent days. To avoid
this, if the KBDI value for the previous day is missing, the function uses
the most recent non-missing value from the previous 30 days. If no
non-missing value is available, the maximum value of 203.2 is used.
}
