% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdi_functions.R
\name{calculate_ffdi}
\alias{calculate_ffdi}
\title{Calculate FFDI based on weather records for one or more stations}
\usage{
calculate_ffdi(
  dat,
  fields = c(station = "station", date = "date_local", hour = "hour_local", minute =
    "min_local", precipitation = "precipitation", temperature = "temperature",
    relhumidity = "relhumidity", windspeed = "windspeed"),
  av.rainfall = NULL,
  datatype = c("guess", "aws", "synoptic")
)
}
\arguments{
\item{dat}{A data frame with records for one or more weather stations
with columns for station number, date, hour, minute, precipitation,
temperature, relative humidity and wind speed. See \code{fields}
argument (below) for how to provide non-default column names.}

\item{fields}{A named vector or list where names are variables (station,
date, hour, minute, precipitation, temperature, relhumidity, windspeed) and
values are the correspdonding column names in the input data frame.}

\item{av.rainfall}{Average annual rainfall value(s) to use in the calculation
of KBDI (on which FFDI relies). Can either be a single numeric value to be
applied to all stations in the input data, or a two-column matrix or data
frame with station numbers in the first column and average rainfall values
in the second column.}

\item{datatype}{Either a specified data type ('aws' or 'synoptic') or 'guess'
(default) to guess the type from the data. Case-insensitive and may be
abbreviated. The data type controls how rainfall values are aggregated to
daily values for the calculation of KBDI and drought factor. AWS rainfall
values are total since 9am, whereas Synoptic values are rainfall for the
record time step. If 'guess', the function checks whether there are records
for sub-hourly time steps (AWS data) or not (Synoptic data).}
}
\value{
A data frame with columns:
  \code{'station'};
  date, hour and minute (named to match input names, e.g. date_local);
  \code{'tmaxdaily'} (maximum temperature for calendar day);
  \code{'precipdaily'} (total rainfall from 09:01 to 09:00 the next day);
  \code{'kbdi'} (Keetch-Bryam drought index);
  \code{'drought'} (daily drought factor values);
  \code{'ffdi'} (FFDI value based on daily drought factor and time-step
    values for temperature, wind speed and relative humidity.
}
\description{
This function calculates FFDI (Forest Fire Danger Index) and the related
variables: KBDI (Keetch-Bryam Drought Index) and Drought Factor. If there are
missing days in the time series, dummy records will be added with all weather
variables set to \code{NA}. This allows KBDI, drought factor and FFDI to be
calculated for the rest of the time series. FFDI values for the dummy records
will be set to \code{NA}. KBDI values (and resulting drought factor values)
are set following the protocol described for \code{\link{calculate_kbdi}}. The
records for each station must cover a period of at least 21 days for FFDI to
be calculated, since Drought Factor requires an initial 20 days of daily
rainfall data.
}
